pyproject.toml
[project]
name = "lmi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.1.0",
    "pluggy>=1.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "rich>=13.0.0",
    "authlib>=1.3.0",
    "keyring>=24.3.1",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "coverage>=7.8.0",
    "pytest>=8.3.5",
]
docs = [
    "mkdocs>=1.6.1",
    "mkdocs-terminal>=1.3.0",
]

[tool.black]
line-length = 88

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "I", "N", "D", "UP", "B", "C4", "T20", "A", "COM", "DTZ", "EM", "ISC", "ICN", "PIE", "PT", "Q", "RET", "SIM", "TID", "TCH", "ARG", "PTH", "ERA", "PL", "RUF"]
ignore = []
exclude = [".venv", "__pycache__", ".git", ".ruff_cache"]

[project.scripts]
lmi = "lmi.__main__:cli"

[project.entry-points.lmi_plugins]
# Example: 'my_plugin = my_package.my_plugin:PluginClass'
# Plugins can be added here by external packages or for testing
